Update Kotlin version and configure tests in build.gradle.kts
Downgraded the Kotlin version from 1.9.10 to 1.8.20 to ensure compatibility with certain systems. New dependencies, gradleTestKit() and junit-jupiter, were added for test implementation, and configuration was made for these tests to use JUnit Platform. The GreetingTask was moved to the org.gradle package and was refactored for more concise property definitions. Tasks were also uncommented in build.gradle.kts to enable the GreetingTask to function, providing a greeting to the user.

These changes are necessary for increased efficiency and compatibility, as well as for improved test functionality.